Fix thumbnails being garbled, mostly.
This is a mix of a couple approaches:

* Instead of using QPixmap::scaled, using QPixmap::transform and
  specifying a smooth transformation.

* Increasing size to 100x100, originally 70x70.

This has been on my mind for some time. I still think it's possible to
have even nicer thumbnails like what ristretto has. However, given that
this is a small personal project, and I suspect there will be
diminishing returns, I'm going to call this good enough and leave it as
is.